second commit

+++ src/main/java/controles/DeleteIncident.java
@@ -4,7 +4,9 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
+import entities.AlertHelper;
 import entities.Incident;
+import entities.Injury;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -13,6 +15,7 @@ import javafx.scene.Parent;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.ComboBox;
+import javafx.stage.Stage;
 import services.IncidentServices;
 
 public class DeleteIncident implements Initializable {
@@ -28,6 +31,8 @@ public class DeleteIncident implements Initializable {
 
     @FXML
     private Button DeleteIncident;
+    @FXML
+    private Button back;
 
     @FXML
     void initialize() {
@@ -35,6 +40,16 @@ public class DeleteIncident implements Initializable {
     }
     @FXML
     void DeleteIncident(ActionEvent event) {
+        Incident selectedInc = IncidentToDelete.getValue();
+
+        if (selectedInc == null) {
+            // No injury selected, show an error message
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setContentText("Please select an incident to delete.");
+            alert.show();
+            return;
+        }
+
         IncidentServices incidentServices = new IncidentServices();
         try {
             incidentServices.deleteEntity(IncidentToDelete.getValue());
@@ -55,5 +70,26 @@ public class DeleteIncident implements Initializable {
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
         IncidentToDelete.getItems().addAll(new IncidentServices().getAllData());
+        back.setOnMouseClicked(event -> {
+            Stage primaryStage = (Stage) back.getScene().getWindow(); // Get primaryStage
+            Home.switchScene(primaryStage, "/displayIncident.fxml");
+        });
+
+        DeleteIncident.setOnAction(event -> {
+            String confirmationMessage = "Are you sure you want to delete this incident?";
+            boolean confirmed = AlertHelper.showConfirmationAlert(confirmationMessage);
+            if (confirmed) {
+                // Perform the delete operation
+                // For example: deleteIncident();
+                Alert alert = new Alert(Alert.AlertType.INFORMATION);
+                alert.setContentText("Incident deleted successfully");
+                alert.show();
+
+            } else {
+                Alert alert2 = new Alert(Alert.AlertType.ERROR);
+                alert2.setContentText("Delete operation cancelled.");
+                alert2.show();
+            }
+        });
     }
 }
